Add label:get command and tests
Added a new command, "label:get", in the application which is responsible for displaying a list of translations for a specific label. Additionally, unit tests are created to ensure the proper functioning of this new command. Also made minor changes to the README file to reflect these changes and updated the application's description and version in package.json.